The provided ruby code and accompanying tests won't pass for recent versions of Ruby.
```ruby
$ ruby test_yatzy.rb
test_yatzy.rb:1:in `require_relative': --> ~/Yatzy-Refactoring-Kata/ruby/yatzy.rb
_5 is reserved for numbered parameters
    1  class Yatzy
>  86    def initialize(d1, d2, d3, d4, _5)
>  93    end
  256  end
.... # similar errors elided for brevity
```
This is because the provided file follows the convention of the original problem, and uses underscore number in variables names: `_1`, `_2` etc.

Since Ruby 2.7 (released 2019), [numbered parameters have been reserved](https://bugs.ruby-lang.org/issues/4475) in the language as default block parameters.

This means that, in order to get passing tests to start the refactor, in currently supported versions of Ruby, you need to rename these variables.

To fix this, we rename the `_1` to `_one` and so on. Still inconsistant,
though less likely to have been written accidentally. Also: still crying
out for refactoring, and the tests still pass.

This also has the benefit of keeping the spirit of the original alive.